leaveWorld.ts - conditional check to remove homebase cell only if one is already present

+++ server/src/services/maproom/v2/leaveWorld.ts
@@ -34,5 +34,5 @@ const removeUserCells = async (user: User) => {
   });
 
   if (outposts.length > 0) await ORMContext.em.removeAndFlush(outposts);
-  await ORMContext.em.removeAndFlush(homeBase);
+  if (homeBase) await ORMContext.em.removeAndFlush(homeBase);
 };
